\name{DT_fulldiallel}
\alias{DT_fulldiallel}
\docType{data}
\title{
Full diallel data for corn hybrids
}
\description{
This dataset contains phenotpic data for 36 winter bean hybrids, coming from a full diallel design and evaluated for 9 traits. The column male and female origin columns are included as well. 

}
\usage{data("DT_fulldiallel")}
\format{
  The format is:
 chr "DT_fulldiallel"
}
\source{
This data was generated by a winter bean study and originally included in the agridat package.
}
\references{

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommer. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744 

}
\examples{

# ####=========================================####
# #### For CRAN time limitations most lines in the
# #### examples are silenced with one '#' mark,
# #### remove them and run the examples
# ####=========================================####
 data(DT_fulldiallel)
# DT <- DT_fulldiallel
# head(DT)
# mix <- lmebreed(stems~1 + (1|female)+(1|male), data=DT)
# vc <- VarCorr(mix); print(vc,comp=c("Variance"))

}
\keyword{datasets}
\seealso{The core function of the package \code{\link{lmebreed}} }
