\name{DT_h2}
\alias{DT_h2}
\docType{data}
\title{
Broad sense heritability calculation.
}
\description{
This dataset contains phenotpic data for 41 potato lines evaluated in 5 locations across 3 years in an RCBD design. The phenotypic trait is tuber quality and we show how to obtain an estimate of DT_h2 for the trait.

}
\usage{data("DT_h2")}
\format{
  The format is:
 chr "DT_h2"
}
\source{
This data was generated by a potato study.
}
\references{

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommer. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744 

}
\examples{
# 
# ####=========================================####
# #### For CRAN time limitations most lines in the 
# #### examples are silenced with one '#' mark, 
# #### remove them and run the examples
# ####=========================================####
 data(DT_h2)
# DT <- DT_h2
# head(DT)
# ####=========================================####
# #### fit the mixed model (very heavy model)
# ####=========================================####
# DT=DT[with(DT, order(Env)), ]
# 
# Z <- with(DT, dsc(Env))$Z
# for(i in 1:ncol(Z)){DT[,colnames(Z)[i]] <- Z[,i]}
# csdiagFormula <- paste0( "y ~ Env + (", paste(colnames(Z), collapse = "+"), "|| Name)")
# blockFormula <- paste0( "(0+", paste(colnames(Z), collapse = "+"), "|| Block)")
# form <- paste(csdiagFormula , blockFormula, sep="+")
# 
# ans1b <- lmebreed(as.formula(form),
#                   data=DT)
# vc <- VarCorr(ans1b); print(vc,comp=c("Variance"))


}
\keyword{datasets}
\seealso{The core function of the package \code{\link{lmebreed}} }
