\name{DT_halfdiallel}
\alias{DT_halfdiallel}
\docType{data}
\title{
half diallel data for corn hybrids
}
\description{
This dataset contains phenotpic data for 21 corn hybrids, with 2 technical repetitions, coming from a half diallel design and evaluated for sugar content. The column geno indicates the hybrid and male and female origin columns are included as well.

}
\usage{data("DT_halfdiallel")}
\format{
  The format is:
 chr "DT_halfdiallel"
}
\source{
This data was generated by a corn study.
}
\references{

Covarrubias-Pazaran G (2016) Genome assisted prediction of quantitative traits using the R package sommer. PLoS ONE 11(6): doi:10.1371/journal.pone.0156744 

}
\examples{

data("DT_halfdiallel")
# DT <- DT_halfdiallel
# head(DT)
# DT$femalef <- as.factor(DT$female)
# DT$malef <- as.factor(DT$male)
# DT$genof <- as.factor(DT$geno)
# # overlay matrix to be added to the addmat argument
# Z <- with(DT, overlay(femalef,malef) )
# Z <- Z[which(!is.na(DT$sugar)),]
# # inital values for incidence matrix but irrelevant
# # since these will be replaced by admat argument
# fema <- (rep(colnames(Z), nrow(DT)))[1:nrow(DT)]
# #### model using overlay without relationship matrix
# modh <- lmebreed(sugar ~ (1|genof) + (1|fema),
#                  addmat = list(fema=Z),
#              data=DT)
# vc <- VarCorr(modh); print(vc,comp=c("Variance"))
# #### model using overlay with relationship matrix
# # relationship matrix to be added to the relmat argument
# A <- diag(7); colnames(A) <- rownames(A) <- 1:7;A
# modh <- lmebreed(sugar ~ (1|genof) + (1|fema),
#                  addmat = list(fema=Z),
#                  relmat = list(fema=Matrix::chol(A)),
#                  data=DT)
# vc <- VarCorr(modh); print(vc,comp=c("Variance"))


}
\keyword{datasets}
\seealso{The core function of the package \code{\link{lmebreed}} }
