\name{simage}
\alias{simage}
\title{Image of sparsity between two variables}
\description{\code{image} for sparsity.
}
\usage{
simage(data, \dots)
}
\arguments{
 \item{data}{model data of class \code{"data.frame"}}
 \item{Var1}{variable to set in the x axis}
 \item{Var2}{variable to set in the y axis}
 \item{\dots}{Further arguments to be passed to the image function.}
}
\value{
  vector of image
 }
\author{Giovanny Covarrubias}
\seealso{\code{\link{image}}, \code{\link{lmebreed}}}
\keyword{models}

